refactor: migrate dashboard and task list to use fetch command hooks
